Add basic Terraform Acceptance Test (#9)
This adds a simple, example Terraform Provider Acceptance test. Details
on Hashicorp's acceptance testing framework are at:
https://www.terraform.io/docs/extend/testing/acceptance-tests/index.html

To run the tests: set your APTIBLE_AUTH_ROOT_URL, APTIBLE_API_ROOT_URL,
and APTIBLE_ACCESS_TOKEN environment variables. Then run `make testacc`

It will check for go fmt conformance, so be sure to run that first if
needed.

Right now the environment (aka Account) ID is hardcoded and will likely
need to be changed by each of us to match whichever account we're using
in our sandboxes. Eventually, we'll want to standardize that in its own
test sandbox, but that was bogging this down.

This also includes some re-arranging of things to better match what the
acceptance test framework assumes. The Makefile is taken from
Hashicorp's demo project so not all of the targets there make sense.