Add field to capture the endpoint's domain
The virtual domain field contains the appropriate URL to access an
endpoint.

I decided to call it `endpoint_domain` here because that seemed clearer
to me than `virtual_domain`. I couldn't find any references in either
the dashboard or the docs to that domain, so I made a choice. If there's
something else out there that already refers to that domain, then we
should change here for consistency.

We'll probably also want to add this to our Terraform docs.