Add all database credentials as fields on the database resource (#28)
Previously, we only set the default credentials as the `connection_url`
property. This did not work well for Redis databases where the secured
credentials were in a second record.

Now the default is in `default_connection_url` renamed from
`connection_url` and all of the credentials are available in an array
called `connection_urls`. The first item in that array should be the
default.

There is a bug in our generated client that does not include the
credentials in the original database object. To work around that, we
make a separate API call to get the credentials here. Once we have fixed
the generated client, we can remove that separate call.