Fix connection_url acceptance tests (#29)
When I changed the `connection_url` to `default_connection_url` to
allow for finding multiple credentials, I did not update all of the
tests. This fixes those tests to refer to the correct value.

Also, I updated replicas to use the new field name for consistency. I
did *not* update them to return multiple credentials yet. I plan to
update replicas to do that once we have the "real" fix to correct the
generated client.