Update endpoint documentation to use process_type (#33)
Closes: https://github.com/aptible/terraform-provider-aptible/issues/32

Our documentation referred to a `service_name` field on the endpoints,
but the field is actually `process_type`. This updates the documentation
to accurately reflect the current usage.

If we decide to switch to `service_name` in the future, we can always
switch the docs back when we make the code change.