fix(app): schema should drive whether we should crud sizing policy
We were using a boolean `update` to determine whether we should create
or update a service sizing policy associated with a service.  This
boolean was set whether an App was being created or updated.  This is
flawed logic because an App can exist with autoscaling enable, disabled,
or toggled between the two states.  Therefore, we need to be able to
crud a service sizing policy regardless if an App is being created or
updated.

This change removes the boolean `update` and adjust the service sizing
crud based on whether a service sizing policy exists in `deploy-api` or
not.

Reference: https://app.shortcut.com/aptible/story/30748/bug-report-if-a-deployed-app-service-has-autoscaling-disabled-you-cannot-apply-or-update-the-autos